# openapi.yaml
openapi: 3.0.3
info:
  title: LucidLayer API
  version: 0.2.0
  description: |
    LucidLayer Offchain API.

    This OpenAPI spec is the source of truth for the actual backend routes.
    It is used to generate SDK clients via Speakeasy.

    Route groups:
    - `/v1/*` — Passports, Match, Run, Receipts, Epochs, Payouts, Compute
    - `/health/*` — System health and dependency checks
    - `/api/agents/*` — AI Agent MMR, Planner, and Orchestrator
servers:
  - url: https://api.lucid.foundation
    description: Production server

# Speakeasy MCP tool generation configuration
...
paths:
...
  /v1/match/explain:
    post:
      tags: [Match]
      operationId: lucid_match_explain
      summary: Evaluate policy against compute/model meta
      x-speakeasy-mcp:
        name: lucid_match_explain
        description: Evaluate a policy against compute & model metadata and explain the decision.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policy: { $ref: '#/components/schemas/Policy' }
                compute_meta: { $ref: '#/components/schemas/ComputeMeta' }
                model_meta: { $ref: '#/components/schemas/ModelMeta' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, allowed, reasons, policy_hash]
                properties:
                  success: { type: boolean }
                  allowed: { type: boolean }
                  reasons:
                    type: array
                    items: { type: string }
                  policy_hash: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/match:
    post:
      tags: [Match]
      operationId: lucid_match
      summary: Match compute for model
      x-speakeasy-mcp:
        name: lucid_match
        description: Match compute for a model given a compute catalog and policy constraints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model_meta: { $ref: '#/components/schemas/ModelMeta' }
                policy: { $ref: '#/components/schemas/Policy' }
                compute_catalog:
                  type: array
                  items: { $ref: '#/components/schemas/ComputeMeta' }
                require_live_healthy: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, match, explain]
                properties:
                  success: { type: boolean }
                  match: { type: object, additionalProperties: true }
                  explain: { type: object, additionalProperties: true }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/route:
    post:
      tags: [Match]
      operationId: lucid_route
      summary: Plan a route (match + resolve endpoint)
      x-speakeasy-mcp:
        name: lucid_route
        description: Match + resolve an executable inference endpoint for a model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model_meta: { $ref: '#/components/schemas/ModelMeta' }
                policy: { $ref: '#/components/schemas/Policy' }
                compute_catalog:
                  type: array
                  items: { $ref: '#/components/schemas/ComputeMeta' }
                request_id: { type: string }
                require_live_healthy: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, route, explain]
                properties:
                  success: { type: boolean }
                  request_id: { type: string }
                  route:
                    type: object
                    properties:
                      compute_passport_id: { type: string }
                      model_passport_id: { type: string }
                      endpoint: { type: string }
                      runtime: { type: string }
                      policy_hash: { type: string }
                      fallbacks:
                        type: array
                        items: { type: object, additionalProperties: true }
                  explain: { type: object, additionalProperties: true }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ===========================================================================
  # Run / Inference
  # ===========================================================================
  /v1/run/inference:
    post:
      tags: [Run]
      operationId: lucid_run_inference
      summary: Run inference (optionally streaming via SSE)
      x-speakeasy-mcp:
        name: lucid_run_inference
        description: |
          Run inference through LucidLayer.
          NOTE: streaming responses may not be supported by all MCP clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
      responses:
        '200':
          description: Non-streaming inference result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceResult'
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/chat/completions:
    post:
      tags: [Run]
      operationId: lucid_chat_completions
      summary: OpenAI-compatible chat completions
      x-speakeasy-mcp:
        name: lucid_chat_completions
        description: OpenAI-compatible chat completions routed through LucidLayer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ===========================================================================
  # Compute live-state
  # ===========================================================================
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    Policy:
      type: object
      required: [policy_version]
      additionalProperties: false
      properties:
        policy_version: { type: string, description: "Policy schema version, e.g. '1.0'" }
        allow_regions:
          type: array
          items: { type: string }
        residency_required: { type: boolean }
        attestation:
          type: object
          additionalProperties: false
          properties:
            attestation_required: { type: boolean }
            require_cc_on: { type: boolean }
            fallback_allowed: { type: boolean }
        latency:
          type: object
          additionalProperties: false
          properties:
            p95_ms_budget: { type: integer }
            hard_timeout_ms: { type: integer }
        cost:
          type: object
          additionalProperties: false
          properties:
            max_price_per_1k_tokens_usd: { type: number }
            spot_only: { type: boolean }
        privacy:
          type: object
          additionalProperties: false
          properties:
            store_inputs: { type: boolean }
            redact_pii: { type: boolean }

    # =========================================================================
    # ComputeMeta / ModelMeta schemas  (mirror schemas/*.schema.json)
    # =========================================================================
    ComputeHardware:
      type: object
      required: [gpu, vram_gb]
      properties:
        gpu: { type: string, description: "GPU model (e.g. 'NVIDIA-A100-40GB')" }
        vram_gb: { type: integer, minimum: 0, description: GPU VRAM in gigabytes }
        arch: { type: string, description: "GPU architecture (e.g. 'ampere', 'hopper')" }
        gpu_count: { type: integer, minimum: 1, default: 1 }
        cpu_cores: { type: integer, minimum: 1 }
        memory_gb: { type: integer, minimum: 1, description: System RAM in gigabytes }
    ComputeRuntime:
      type: object
      required: [name]
      properties:
        name:
          type: string
          enum: [vllm, tgi, tensorrt, hf-inference-api, openai-compatible, custom]
        version: { type: string }
        image: { type: string, description: Docker image reference }
    ComputeCapabilities:
      type: object
      properties:
        supports_streaming: { type: boolean, default: true }
        supports_attestation: { type: boolean, default: false }
        supports_cc_on: { type: boolean, default: false }
        inference_types:
          type: array
          items:
            type: string
            enum: [text-generation, chat-completion, embeddings, image-generation, speech-to-text]
    ComputeEndpoints:
      type: object
      required: [inference_url]
      properties:
        inference_url: { type: string, description: URL for inference requests }
        quote_url: { type: string, format: uri }
        jobs_url: { type: string, format: uri }
        health_url: { type: string, format: uri }
        metrics_url: { type: string, format: uri }
    ComputeWorker:
      type: object
      required: [worker_id, status]
      properties:
        worker_id: { type: string }
        status:
          type: string
          enum: [online, offline, degraded, draining]
        last_heartbeat: { type: integer, description: Unix timestamp of last heartbeat }
    ComputeMeta:
      type: object
      description: >-
        Metadata for a compute provider passport.
        Validated against schemas/ComputeMeta.schema.json (additionalProperties: false).
      required: [schema_version, compute_passport_id, provider_type, regions, hardware, runtimes, endpoints]
      properties:
        schema_version: { type: string, enum: ['1.0'] }
        compute_passport_id: { type: string, minLength: 10 }
        provider_type:
          type: string
          enum: [depin, cloud, onprem, managed]
        execution_mode:
          type: string
          enum: [byo_runtime, managed_endpoint]
        operator_pubkey:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: ed25519 public key (hex) of the provider operator
        regions:
          type: array
          items: { type: string }
          minItems: 1
        residency_supported: { type: boolean, default: false }
        hardware: { $ref: '#/components/schemas/ComputeHardware' }
        gpu_fingerprint:
          type: string
          nullable: true
          description: "GPU hardware fingerprint. Must be null for managed_endpoint mode."
        runtime_hash:
          type: string
          nullable: true
          pattern: '^sha256:[a-f0-9]{64}$'
          description: "Docker image digest. Must be null for managed_endpoint mode."
        runtimes:
          type: array
          items: { $ref: '#/components/schemas/ComputeRuntime' }
          minItems: 1
        capabilities: { $ref: '#/components/schemas/ComputeCapabilities' }
        network:
          type: object
          properties:
            p95_ms_estimate: { type: integer, minimum: 0 }
            bandwidth: { type: string }
        limits:
          type: object
          properties:
            max_context: { type: integer, minimum: 0 }
            max_batch: { type: integer, minimum: 0 }
            max_input_tokens: { type: integer, minimum: 1 }
            max_output_tokens: { type: integer, minimum: 1 }
            requests_per_minute: { type: integer, minimum: 1 }
        pricing:
          type: object
          properties:
            price_per_1k_tokens_estimate: { type: number, minimum: 0 }
            price_per_minute_estimate: { type: number, minimum: 0 }
            per_input_token: { type: number, minimum: 0 }
            per_output_token: { type: number, minimum: 0 }
            currency:
              type: string
              enum: [lamports, usd_cents, credits]
        endpoints: { $ref: '#/components/schemas/ComputeEndpoints' }
        workers:
          type: array
          items: { $ref: '#/components/schemas/ComputeWorker' }
        sla:
          type: object
          properties:
            tier:
              type: string
              enum: [best-effort, standard, premium, enterprise]
            uptime_target: { type: number, minimum: 0, maximum: 100 }
        policy_tags:
          type: array
          items: { type: string }
          description: "Tags for matching (e.g. 'hipaa', 'gdpr', 'no-logging')"
        metadata:
          type: object
          additionalProperties: true
    ModelRequirements:
      type: object
      properties:
        min_vram_gb: { type: integer, minimum: 0 }
        gpu_classes:
          type: array
          items: { type: string }
        cuda_min: { type: string }
    ModelMeta:
      type: object
      description: >-
        Metadata for a model passport.
        Validated against schemas/ModelMeta.schema.json (additionalProperties: false).
      required: [schema_version, model_passport_id, format, runtime_recommended]
      properties:
        schema_version: { type: string, enum: ['1.0'] }
        model_passport_id: { type: string, minLength: 10 }
        format:
          type: string
          enum: [safetensors, gguf, api]
        runtime_recommended:
          type: string
          enum: [vllm, tgi, tensorrt, trustgate, openai]
        name: { type: string }
        description: { type: string }
        provider: { type: string, description: "Provider name (e.g. trustgate, openai)" }
        base:
          type: string
          enum: [hf, openai, anthropic, google, cohere, custom_endpoint]
          default: hf
        hf:
          type: object
          properties:
            repo_id: { type: string }
            revision: { type: string }
        context_length: { type: integer, minimum: 1 }
        quantizations:
          type: array
          items:
            type: string
            enum: [fp16, int8, awq, gptq, gguf_q4]
        requirements: { $ref: '#/components/schemas/ModelRequirements' }
        endpoints:
          type: object
          properties:
            openai_compat_base_url: { type: string }
            auth_mode:
              type: string
              enum: [byok, lucid_key, none]
              default: none
        artifacts:
          type: array
          items: { type: string }
        weights_uri: { type: string }
        weights_hash: { type: string }

    # =========================================================================
    # Chat / Inference schemas
    # =========================================================================
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, function]
        content: { type: string }
        name: { type: string }
        function_call:
          type: object
          properties:
            name: { type: string }
            arguments: { type: string }
    InferenceRequest:
      type: object
      properties:
        model_passport_id: { type: string }
        model: { type: string }
        prompt: { type: string }
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
        max_tokens: { type: integer }
        temperature: { type: number }
        top_p: { type: number }
        top_k: { type: integer }
        stop:
          type: array
          items: { type: string }
        stream: { type: boolean }
        policy: { $ref: '#/components/schemas/Policy' }
        compute_passport_id: { type: string }
        trace_id: { type: string }
        request_id: { type: string }
    InferenceResult:
      type: object
      required: [success, run_id, tokens_in, tokens_out, ttft_ms, total_latency_ms, model_passport_id, compute_passport_id, runtime]
      properties:
        success: { type: boolean }
        run_id: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
        text: { type: string }
        finish_reason: { type: string }
        tokens_in: { type: integer }
        tokens_out: { type: integer }
        ttft_ms: { type: integer }
        total_latency_ms: { type: integer }
        model_passport_id: { type: string }
        compute_passport_id: { type: string }
        runtime: { type: string }
        policy_hash: { type: string }
        receipt_id: { type: string }
        used_fallback: { type: boolean }
        fallback_reason: { type: string }
        error: { type: string }
        error_code: { type: string }
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
        max_tokens: { type: integer }
        temperature: { type: number }
        top_p: { type: number }
        stop:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        stream: { type: boolean }
        policy: { $ref: '#/components/schemas/Policy' }
        trace_id: { type: string }
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id: { type: string }
        object: { type: string }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { $ref: '#/components/schemas/ChatMessage' }
              finish_reason: { type: string, nullable: true }
        usage:
          type: object
          required: [prompt_tokens, completion_tokens, total_tokens]
          properties:
            prompt_tokens: { type: integer }
            completion_tokens: { type: integer }
            total_tokens: { type: integer }

    # =========================================================================
    # Error
    # =========================================================================
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean }
        error: { type: string }
        message: { type: string }
        error_code: { type: string }
        details: {}

    # =========================================================================
    # Compute
    # =========================================================================
...